feat: implement AuthService and AuthController for user synchronization
