+++ src/contexts/CoreContext/application/services/UserService.ts
@@ -6,7 +6,6 @@ import { CreateUserFreelancerProfileUseCase } from "../useCases/CreateUserFreela
 import { CreateUserUseCase } from "../useCases/CreateUserUseCase";
 import { GetUserProfileUseCase } from "../useCases/GetUserProfileUseCase";
 import { GetUserUseCase } from "../useCases/GetUserUseCase";
-import IUseCase from "@/contexts/LearningContext/domain/interfaces/IUseCase";
 
 @injectable()
 export class UserService implements IUserService {
@@ -18,9 +17,7 @@ export class UserService implements IUserService {
     @inject("CreateUserFreelancerProfileUseCase")
     private readonly createUserFreelancerProfileUseCase: CreateUserFreelancerProfileUseCase,
     @inject("GetUserProfileUseCase")
-    private readonly getUserProfileUseCase: GetUserProfileUseCase,
-    @inject("SyncUserUseCase")
-    private readonly syncUserUseCase: IUseCase<User, User>
+    private readonly getUserProfileUseCase: GetUserProfileUseCase
   ) {}
   async createFreelanceProfile(id: string): Promise<User> {
     const updatedUser =
@@ -48,9 +45,5 @@ export class UserService implements IUserService {
   }
   async getClientProfile(userId: string): Promise<IGetUserProfileDto> {
     return await this.getUserProfileUseCase.execute(userId);
-  }
-
-  async syncUser(user: User): Promise<User> {
-    return await this.syncUserUseCase.execute(user);
   }
 }
