+++ src/contexts/CoreContext/domain/interfaces/controllers/IAuthController.ts
@@ -1,0 +1,5 @@
+import { Request, Response } from "express";
+
+export interface IAuthController {
+  syncUser(req: Request, res: Response): Promise<void>;
+}
