+++ src/contexts/CoreContext/domain/interfaces/controllers/IUserController.ts
@@ -5,5 +5,4 @@ export interface IUserController {
   post(req: Request, res: Response): void;
   freelance(req: Request, res: Response): void;
   getUserProfile(req: Request, res: Response): Promise<void>;
-  syncUser(req: Request, res: Response): Promise<void>;
 }
