+++ src/contexts/CoreContext/domain/interfaces/services/IAuthService.ts
@@ -1,0 +1,5 @@
+import { User } from "../../aggregates/User";
+
+export interface IAuthService {
+  syncUser(user: User): Promise<User>;
+}
